requirements.txt
# Zerodha Kite Connect
kiteconnect==5.0.1

# Core dependencies
requests>=2.28.0
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
ta-lib>=0.4.28
yfinance>=0.2.28

# Database
psycopg2-binary>=2.9.0
redis>=4.5.0
sqlalchemy>=2.0.0

# API & Webhooks
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Data processing
scipy>=1.10.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
schedule>=1.2.0

# Monitoring & Alerts
prometheus-client>=0.17.0
python-telegram-bot>=20.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
backtrader>=1.9.78.123
